refactor(decomp): extract audit correlation + cross_tenant_guard → new crate (TD-DECOMP-16)
Creates crates/horizontal/proximadb-audit. Moves correlation.rs (29t),
cross_tenant_guard.rs (12t), types.rs (0t) = 41 tests. AuditData +
collect_comprehensive_audit_data moved to crate (inherent-impl fix).
Root mod.rs re-exports from crate; logger.rs + storage.rs stay root-side.

Verified: audit crate 41 tests pass; root check + server build + fmt green.
