+++ Cargo.toml
@@ -66,6 +66,7 @@ members = [
     "crates/horizontal/proximadb-recall-bench",
     "crates/horizontal/proximadb-resilience",
     "crates/horizontal/proximadb-automl",
+    "crates/horizontal/proximadb-audit",
     "crates/horizontal/proximadb-runtime-common",
     "crates/horizontal/proximadb-serialization",
     "crates/horizontal/proximadb-security",
@@ -268,6 +269,7 @@ proximadb-serialization = { path = "crates/horizontal/proximadb-serialization" }
 proximadb-tls = { path = "crates/horizontal/proximadb-tls" }
 proximadb-resilience = { path = "crates/horizontal/proximadb-resilience" }
 proximadb-automl = { path = "crates/horizontal/proximadb-automl" }
+proximadb-audit = { path = "crates/horizontal/proximadb-audit" }
 proximadb-storage-common = { path = "crates/storage/proximadb-storage-common" }
 proximadb-storage-ports = { path = "crates/storage/proximadb-storage-ports" }
 proximadb-storage-traits = { path = "crates/storage/proximadb-storage-traits" }
@@ -385,6 +387,7 @@ proximadb-security = { path = "crates/horizontal/proximadb-security" }
 proximadb-tls = { path = "crates/horizontal/proximadb-tls" }
 proximadb-resilience = { workspace = true }
 proximadb-automl = { workspace = true }
+proximadb-audit = { workspace = true }
 proximadb-storage-common = { path = "crates/storage/proximadb-storage-common" }
 proximadb-storage-ports = { path = "crates/storage/proximadb-storage-ports" }
 proximadb-storage-traits = { path = "crates/storage/proximadb-storage-traits" }
