+++ crates/horizontal/proximadb-audit/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "proximadb-audit"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "Audit correlation analysis + cross-tenant guard — extracted from root (TD-DECOMP-16)"
+
+[lib]
+name = "proximadb_audit"
+path = "src/lib.rs"
+
+[dependencies]
+anyhow.workspace = true
+dashmap.workspace = true
+chrono.workspace = true
+proximadb-security = { path = "../../horizontal/proximadb-security" }
+serde.workspace = true
+serde_json.workspace = true
+tokio.workspace = true
+tracing.workspace = true
