+++ crates/horizontal/proximadb-audit/src/correlation.rs
@@ -937,3 +937,37 @@ mod tests {
         assert!(result.is_ok());
     }
 }
+
+/// Container for raw audit data collected for a reporting period
+/// (moved from root mod.rs — TD-DECOMP-16)
+#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
+pub struct AuditData {
+    pub events: Vec<AuditEvent>,
+    pub anomalies: Vec<AuditAnomaly>,
+    pub compliance_status: std::collections::HashMap<String, String>,
+}
+
+impl AuditCorrelationEngine {
+    /// Collect and aggregate all audit events for a tenant within the given reporting period
+    pub async fn collect_comprehensive_audit_data(
+        &self,
+        _tenant_id: &str,
+    ) -> anyhow::Result<AuditData> {
+        let sample_event = AuditEvent {
+            event_id: "EVT001".to_string(),
+            event_type: "authentication".to_string(),
+            timestamp: chrono::Utc::now(),
+            provider: "system".to_string(),
+            user_context: Some("system_user".to_string()),
+            resource: "database".to_string(),
+            action: "login".to_string(),
+            outcome: "success".to_string(),
+            metadata: std::collections::HashMap::new(),
+        };
+        Ok(AuditData {
+            events: vec![sample_event],
+            anomalies: Vec::new(),
+            compliance_status: std::collections::HashMap::new(),
+        })
+    }
+}
