+++ crates/horizontal/proximadb-audit/src/cross_tenant_guard.rs
