+++ crates/horizontal/proximadb-audit/src/types.rs
