+++ docs/10-quality/td/TD-DECOMP-16-extract-audit.adoc
@@ -1,0 +1,29 @@
+// Copyright (C) 2025 ProximaDB
+// SPDX-License-Identifier: Apache-2.0
+= TD-DECOMP-16: Extract audit correlation + cross_tenant_guard → proximadb-audit
+
+[cols="1,3", options="header"]
+|===
+| Field | Value
+| Status | Landing
+| Severity | Low — behavior-neutral re-export
+| Component | new `crates/horizontal/proximadb-audit`; root `src/audit/mod.rs`
+| Relates to | [[TD-DECOMP-1]] (recipe); [[TD-DECOMP-14..15]] (horizontal pattern)
+|===
+
+== Why
+
+`src/audit/{correlation,cross_tenant_guard,types}.rs` (**41 tests**) are fully standalone —
+zero `crate::` + zero `super::` refs. `logger.rs` + `storage.rs` stay root-side (depend on
+root storage/auth). Moved to a new `proximadb-audit` horizontal crate.
+
+== What moved (behavior-neutral)
+
+* New `crates/horizontal/proximadb-audit/`. Moved: `correlation.rs` (29t),
+  `cross_tenant_guard.rs` (12t), `types.rs` (0t) = **41 tests**.
+* Root `src/audit/mod.rs`: re-exports from crate; keeps `logger.rs` + `storage.rs`.
+* Deps: anyhow, chrono, dashmap, serde, serde_json, tokio, tracing, proximadb-security (path).
+
+== Follow-ups
+
+* **Running total once merged: ~582 tests out of root** (541 + 41).
